create post modal duplicate fixed
