+++ apps/admin/pages/news/recommended.tsx
@@ -1,23 +1,9 @@
-import { useState } from 'react'
-
-import {
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalHeader,
-  ModalOverlay,
-  SimpleGrid,
-  useDisclosure,
-} from '@chakra-ui/react'
+import { SimpleGrid } from '@chakra-ui/react'
 import { useSearchModel } from '@wsvvrijheid/services'
-import { Post, RecommendedTopic, StrapiLocale } from '@wsvvrijheid/types'
+import { RecommendedTopic, StrapiLocale } from '@wsvvrijheid/types'
 import {
   AdminLayout,
-  ModelCreateForm,
   ModelCreateModal,
-  postFields,
-  postSchema,
   TopicCard,
   topicFields,
   topicSchema,
@@ -26,35 +12,11 @@ import { useRouter } from 'next/router'
 
 const NewsBookmarkedPage = () => {
   const { locale } = useRouter()
-  const [topic, setTopic] = useState<RecommendedTopic>()
-  const [isNews, setOnNews] = useState<boolean>(false)
 
   const { data, refetch } = useSearchModel<RecommendedTopic>({
     url: 'api/recommended-topics',
     locale: locale as StrapiLocale,
   })
-  const formDisclosure = useDisclosure()
-
-  const handleCreatePost = async (data: RecommendedTopic) => {
-    setTopic(data)
-    setOnNews(true)
-    formDisclosure.onOpen()
-  }
-
-  const postContent = {
-    title: topic?.title,
-    description: topic?.description,
-    content: topic?.description,
-    reference: topic?.url,
-    image: {
-      url: topic?.image,
-    },
-  } as Post
-  const handleSuccess = () => {
-    formDisclosure.onClose()
-    setOnNews(false)
-  }
-
   return (
     <AdminLayout
       title=" Recomended News"
@@ -64,31 +26,6 @@ const NewsBookmarkedPage = () => {
         searchPlaceHolder: 'Search bookmarks',
       }}
     >
-      {isNews && (
-        <Modal
-          isCentered
-          closeOnOverlayClick={true}
-          isOpen={formDisclosure.isOpen}
-          onClose={formDisclosure.onClose}
-          size="6xl"
-          scrollBehavior="inside"
-        >
-          <ModalOverlay />
-          <ModalContent>
-            <ModalHeader color={'primary.500'}>Create Post</ModalHeader>
-            <ModalCloseButton />
-            <ModalBody pos="relative" py={6}>
-              <ModelCreateForm<Post>
-                url="api/posts"
-                schema={postSchema}
-                fields={postFields}
-                model={postContent}
-                onSuccess={handleSuccess}
-              />
-            </ModalBody>
-          </ModalContent>
-        </Modal>
-      )}
       <ModelCreateModal<RecommendedTopic>
         title="Create News"
         url="api/recommended-topics"
@@ -103,7 +40,6 @@ const NewsBookmarkedPage = () => {
           <TopicCard
             key={topic.url}
             topic={{ ...topic, isRecommended: true }}
-            onCreatePost={handleCreatePost}
           />
         ))}
       </SimpleGrid>
