+++ libs/ui/src/admin/ModelForm/ModelCreateModal.tsx
@@ -24,6 +24,7 @@ export const ModelCreateModal = <T extends StrapiModel>({
   children,
   title,
   model,
+  buttonProps,
 }: PropsWithChildren<ModelCreateFormProps<T> & { title: string }>) => {
   const formDisclosure = useDisclosure()
 
@@ -38,6 +39,7 @@ export const ModelCreateModal = <T extends StrapiModel>({
         leftIcon={<IoMdAdd />}
         colorScheme="primary"
         onClick={formDisclosure.onOpen}
+        {...buttonProps}
       >
         {children}
       </Button>
